requirements.txt
# Content & Image Generation MCP Server Dependencies
# For pip users who prefer requirements.txt over pyproject.toml

# Core Framework
fastmcp>=2.13.0

# Google Gen AI SDK (unified SDK for Imagen, Veo, Gemini)
google-genai>=1.0.0

# Anthropic Claude (optional - for content generation)
anthropic>=0.40.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Development Dependencies (optional)
# pytest>=7.0.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# ruff>=0.1.0
